requirements.txt
gitpython==3.1.40
tqdm==4.66.1
transformers==4.38.1
